Add `UnsupportedUserLocationException` for better error handling
Instead of relying on the generic message to notify users that they
are trying to access the API from an unsupported location, this PR
adds an specific new exception that app developers can use.

Although the underlying mechanism is still string matching,
abstracting this into the SDK rather than forcing every developer to
do it on their own is still a win.