Fix API warning task (#138)
Per [b/338644141](https://b.corp.google.com/issues/338644141),

This fixes the `warnAboutAPIChanges` task that was failing when one of
the projects didn't have any changes. This was due to an oversight in
how Gradle propagates skipped properties. We've ran into this before,
but it's easy to forget that even if a task is skipped- Gradle still
considers outputs to be valid (even if they don't exist). So mapping
against it doesn't provide the expected behavior. You can read more
about this on their repo:
[/gradle/issues/7442](https://github.com/gradle/gradle/issues/7442)

This PR also fixes the following:
- [b/338643118](https://b.corp.google.com/issues/338643118) -> Allow
release tasks to work without files existing